TECH-4370: refactor out an abstract monitor
